trying to make code climate happy
